CF-179 CF-177 made changes dropdown does deselect, next button only displays after selection and remove term planner in degree wizard
