+++ frontend/src/pages/DegreeWizard/steps/YearStep.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import { Typography, Button, DatePicker } from "antd";
 import { useDispatch } from "react-redux";
 import "./steps.less";
@@ -11,8 +11,10 @@ const { RangePicker } = DatePicker;
 
 const YearStep = ({ incrementStep, currStep }) => {
   const dispatch = useDispatch();
+  const [nextStep, setNextStep] = useState(false);
 
   const handleYearChange = (_, [startYear, endYear]) => {
+    setNextStep(startYear && endYear);
     const numYears = endYear - startYear + 1;
     dispatch(plannerActions("SET_DEGREE_LENGTH", numYears));
     dispatch(plannerActions("UPDATE_START_YEAR", startYear));
@@ -24,7 +26,7 @@ const YearStep = ({ incrementStep, currStep }) => {
         <Title level={4} className="text">
           What years do you start and finish?
         </Title>
-        {currStep === 1 && (
+        {nextStep && currStep === 1 && (
         <Button type="primary" onClick={incrementStep}>
           Next
         </Button>
