+++ frontend/src/pages/DegreeWizard/steps/steps.less
@@ -14,6 +14,11 @@
   justify-content: space-between;
 }
 
+.steps-start-browsing-container {
+  display: flex;
+  justify-content: flex-end;
+}
+
 .step-duration {
   margin-top: 20px;
   display: flex;
@@ -66,34 +71,6 @@
 }
 .steps-input-warning:focus {
   border: #f46e65 solid 1px;
-}
-
-.steps-grid-cont {
-  display: grid;
-  grid-template-columns: 10em 1fr 1fr 1fr 1fr;
-  align-items: center;
-  max-width: 100em;
-  margin: auto auto;
-  margin-top: 3em;
-}
-
-.steps-grid-item {
-  font-size: 1.5rem;
-  text-align: center;
-}
-
-.steps-term {
-  position: relative;
-}
-
-.steps-term-edit-btn {
-  position: absolute !important;
-  top: 1em;
-  right: 1em;
-}
-
-.steps-term-box {
-  margin-top: 3rem;
 }
 
 .light {
